requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
facenet-pytorch==2.6.0
filelock==3.19.1
fsspec==2025.9.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.12.0.88
pillow==10.2.0
requests==2.32.5
sympy==1.14.0
torch==2.9.0+cpu
torchvision==0.24.0+cpu
tqdm==4.67.1
triton==3.5.0
typing_extensions==4.15.0
urllib3==2.5.0
